compute-logger-0323/function-source/requirements.txt
google-cloud-logging
flask